commit after validation to customer registration and customer login by prathmesh futane

+++ src/app/register/register.component.ts
@@ -1,7 +1,16 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { Customer } from '../appmodel/customer';
-import { CustomerService } from '../customer.service';
+import {
+  Component,
+  OnInit
+} from '@angular/core';
+import {
+  Router
+} from '@angular/router';
+import {
+  Customer
+} from '../appmodel/customer';
+import {
+  CustomerService
+} from '../customer.service';
 
 @Component({
   selector: 'register',
@@ -16,14 +25,19 @@ export class RegisterComponent {
   constructor(private customerService: CustomerService, private router: Router) { }
 
   register() {
-    alert(JSON.stringify(this.customer));
-    this.customerService.register(this.customer).subscribe(response => {
-      alert(JSON.stringify(response));
-      if (response.status == true) {
-        sessionStorage.setItem('customerId', response.registeredCustomerId);
-        this.router.navigate(['login']);
-      }
-    })
+    if (this.customer.password == this.confirmPassword) {
+      alert(JSON.stringify(this.customer));
+      this.customerService.register(this.customer).subscribe(response => {
+        alert(JSON.stringify(response));
+        if (response.status == true) {
+          sessionStorage.setItem('customerId', response.registeredCustomerId);
+          this.router.navigate(['login']);
+        }
+      })
+    } else {
+      alert("password doesnot match")
+    }
+
   }
